+++ packages/manager/apps/pci-databases-analytics/src/data/api/database/service.api.ts
@@ -1,17 +1,17 @@
-import { apiClient } from '@ovh-ux/manager-core-api';
+import {
+  apiClient,
+  createHeaders,
+  IcebergPaginationHeaders,
+  NoCacheHeaders,
+} from '@/data/api/api.client';
 import * as database from '@/types/cloud/project/database';
 import { PCIData, ServiceData } from '.';
 
 export const getServices = async ({ projectId }: PCIData) =>
-  apiClient.v6
-    .get(`/cloud/project/${projectId}/database/service`, {
-      headers: {
-        'X-Pagination-Mode': 'CachedObjectList-Pages',
-        'X-Pagination-Size': '50000',
-        Pragma: 'no-cache',
-      },
-    })
-    .then((res) => res.data as database.Service[]);
+  apiClient.v6.get<database.Service[]>(
+    `/cloud/project/${projectId}/database/service`,
+    { headers: createHeaders(NoCacheHeaders, IcebergPaginationHeaders) },
+  );
 
 export const getService = async <
   T extends database.Service = database.Service
@@ -20,29 +20,34 @@ export const getService = async <
   serviceId,
 }: Omit<ServiceData, 'engine'>) => {
   // first fetch the basic info in services endpoint
-  const serviceInfo = await apiClient.v6
-    .get(`/cloud/project/${projectId}/database/service/${serviceId}`)
-    .then((res) => res.data as database.Service);
+  const serviceInfo = await apiClient.v6.get<database.Service>(
+    `/cloud/project/${projectId}/database/service/${serviceId}`,
+  );
   // then get the engine's specific information
-  const serviceData = await apiClient.v6
-    .get<T>(
-      `/cloud/project/${projectId}/database/${serviceInfo.engine}/${serviceId}`,
-    )
-    .then((res) => res.data);
+  const serviceData = await apiClient.v6.get<T>(
+    `/cloud/project/${projectId}/database/${serviceInfo.engine}/${serviceId}`,
+  );
   return serviceData;
 };
 
 interface AddService extends Omit<ServiceData, 'serviceId'> {
-  serviceInfo: Partial<database.ServiceCreation>;
+  serviceInfo: Omit<
+    Partial<database.ServiceCreation>,
+    'ipRestrictions' | 'disk'
+  > & {
+    ipRestrictions?: Omit<database.service.IpRestriction, 'status'>[];
+    disk?: Partial<database.service.Disk>;
+  };
 }
 export const addService = async ({
   projectId,
   engine,
   serviceInfo,
 }: AddService) =>
-  apiClient.v6
-    .post(`/cloud/project/${projectId}/database/${engine}`, serviceInfo)
-    .then((res) => res.data as database.Service);
+  apiClient.v6.post<database.Service>(
+    `/cloud/project/${projectId}/database/${engine}`,
+    serviceInfo,
+  );
 
 // define which values can be updated in the update service endpoint
 export interface EditService extends ServiceData {
@@ -79,9 +84,10 @@ export const editService = async <
   serviceId,
   data,
 }: EditService) =>
-  apiClient.v6
-    .put<T>(`/cloud/project/${projectId}/database/${engine}/${serviceId}`, data)
-    .then((res) => res.data);
+  apiClient.v6.put<T>(
+    `/cloud/project/${projectId}/database/${engine}/${serviceId}`,
+    data,
+  );
 
 export const deleteService = async ({
   projectId,