+++ packages/manager/apps/pci-databases-analytics/src/pages/services/[serviceId]/backups/_components/BackupsTableColumns.component.tsx
@@ -158,5 +158,5 @@ export const getColumns = ({
     expiryDateColumn,
     statusColumn,
     actionsColumn,
-  ] as ColumnDef<BackupWithExpiricyDate>[];
+  ];
 };