+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/dashboard.controller.js
@@ -1,9 +1,0 @@
-import { PLANS_WITHOUT_BACKUP } from '../../../../../components/project/storages/databases/databases.constants';
-
-export default class DatabasesDashboardCtrl {
-  /* @ngInject */
-  constructor(CHANGELOG) {
-    this.PLANS_WITHOUT_BACKUP = PLANS_WITHOUT_BACKUP;
-    this.CHANGELOG = CHANGELOG;
-  }
-}